ShadowTexture generation works for dir lights

+++ 3D3 Project/src/Engine/RenderView.cpp
@@ -27,16 +27,16 @@ void RenderView::CreateViewport()
 {
 	this->viewport.TopLeftX = 0.0f;
 	this->viewport.TopLeftY = 0.0f;
-	this->viewport.Width = (float)this->width;
-	this->viewport.Height = (float)this->height;
+	this->viewport.Width = this->width;
+	this->viewport.Height = this->height;
 	this->viewport.MinDepth = 0.0f;
 	this->viewport.MaxDepth = 1.0f;
 }
 
 void RenderView::CreateScissorRect()
 {
-	this->scissorRect.left = (long)0;
-	this->scissorRect.right = (long)this->width;
-	this->scissorRect.top = (long)0;
-	this->scissorRect.bottom = (long)this->height;
+	this->scissorRect.left = 0;
+	this->scissorRect.right = this->width;
+	this->scissorRect.top = 0;
+	this->scissorRect.bottom = this->height;
 }
